requirements.txt
streamlit
opencv-python==4.10.0.84
tensorflow
neural-style
pillow
numpy
scipy
Pillow
